requirements.txt
streamlit>=1.28.0
llama-cpp-python>=0.2.0
pydantic>=2.0.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0
numpy>=1.24.0
requests>=2.31.0
duckdb>=0.9.0

